# Faster tests and build improvements