--- a/Makefile
+++ b/Makefile
@@ -1,4 +1,4 @@
-.PHONY: test-clj test-cljs eastwood cljfmt cloverage install smoketest release deploy clean detect_timeout
+.PHONY: test eastwood cljfmt cloverage install smoketest release deploy clean detect_timeout
 
 CLOJURE_VERSION ?= 1.9
 export CLOVERAGE_VERSION = 1.0.13
@@ -12,17 +12,14 @@ JAVA_VERSION = $(shell lein with-profile +sysutils \
 
 source-deps: .source-deps
 
-test-clj: .source-deps smoketest
-	lein with-profile +$(CLOJURE_VERSION),+plugin.mranderson/config,+test-clj test
-
-test-cljs: .source-deps
-	lein with-profile +$(CLOJURE_VERSION),+plugin.mranderson/config,+test-cljs test
+test: .source-deps smoketest
+	lein with-profile +$(CLOJURE_VERSION),+plugin.mranderson/config test
 
 eastwood:
-	lein with-profile +$(CLOJURE_VERSION),+test-clj,+test-cljs,+eastwood eastwood
+	lein with-profile +$(CLOJURE_VERSION),+eastwood eastwood
 
 cljfmt:
-	lein with-profile +$(CLOJURE_VERSION),+test-clj,+test-cljs,+cljfmt cljfmt check
+	lein with-profile +$(CLOJURE_VERSION),+cljfmt cljfmt check
 
 
 # Cloverage can't handle some of the code in this project.  For now we
@@ -31,7 +28,7 @@ cljfmt:
 # exact. See issue #457 for details.
 
 cloverage:
-	lein with-profile +$(CLOJURE_VERSION),+test-clj,+test-cljs,+cloverage cloverage --codecov \
+	lein with-profile +$(CLOJURE_VERSION),+cloverage cloverage --codecov \
 	     -e ".*util.instrument" \
 	     -t "^((?!debug-integration-test).)*$$"