--- a/project.clj
+++ b/project.clj
@@ -119,4 +119,8 @@
                                          if-class [[:block 1]]}}}
 
              :eastwood {:plugins [[jonase/eastwood "0.3.4"]]
-                        :eastwood {:config-files ["eastwood.clj"] :exclude-namespaces [cider-nrepl.plugin cider.tasks cider.nrepl.test-session]}}})
+                        :eastwood {:config-files ["eastwood.clj"]
+                                   :exclude-namespaces [cider-nrepl.plugin
+                                                        cider.tasks
+                                                        cider.nrepl.test-session]
+                                   :namespaces [:source-paths]}}})