--- a/test/clj/cider/nrepl/middleware/debug_integration_test.clj
+++ b/test/clj/cider/nrepl/middleware/debug_integration_test.clj
@@ -2,7 +2,6 @@
   (:require
    [cider.nrepl :refer [wrap-debug]]
    [cider.nrepl.middleware.debug :as d]
-   [cider.nrepl.test.server :refer [start-server]]
    [clojure.test :refer :all]
    [nrepl.core :as nrepl]
    [nrepl.server :as nrepl.server]
@@ -49,7 +48,7 @@
 (defn with-nrepl-session*
   "Run the given function with *transport* and *session-id* bound."
   [f]
-  (with-open [server    (start-server :handler (debug-handler))
+  (with-open [server    (nrepl.server/start-server :handler (debug-handler))
               transport (nrepl/connect :port (:port server))]
     ;; Create a session by sending the "clone" op
     (transport/send transport {:op "clone" :id (next-id)})