… .travis.yml
… test/clj/cider/nrepl/middleware/profile_test.clj
… test/clj/cider/nrepl/middleware/test_test.clj
… test/clj/cider/nrepl/middleware/trace_test.clj
… test/clj/cider/nrepl/middleware/util/meta_test.clj
… test/cljs/cider/nrepl/middleware/cljs_complete_test.clj
… test/cljs/cider/nrepl/middleware/cljs_info_test.clj
… test/cljs/cider/nrepl/middleware/cljs_inspect_test.clj
… test/cljs/cider/nrepl/middleware/cljs_ns_test.clj
… test/cljs/cider/nrepl/middleware/cljs_stacktrace_test.clj
… test/cljs/cider/nrepl/piggieback_test.clj
… test/common/cider/nrepl/test/server.clj
… test/common/cider/nrepl/test_session.clj
… test/common/cider_nrepl/plugin_test.clj
… test/smoketest/project.clj