@bbatsov 2019-01-16
Fantastic work! 🙇 🙇 🙇 

If you have some time, you can do some of that magic to Orchard and the nREPl projects - it'd be nice if they all had consistent build setups.